# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Enki",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/nodes/{node_id}": {
   "patch": {
    "summary": "Update Node",
    "operationId": "update_node_api_nodes__node_id__patch",
    "parameters": [
     {
      "name": "node_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Node Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NodeUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NodeOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Delete Node",
    "operationId": "delete_node_api_nodes__node_id__delete",
    "parameters": [
     {
      "name": "node_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Node Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/analysis/estimate": {
   "get": {
    "summary": "Get Estimate",
    "operationId": "get_estimate_api_analysis_estimate_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Estimate"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/analysis": {
   "post": {
    "summary": "Start Analysis",
    "operationId": "start_analysis_api_analysis_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StartAnalysis"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/topics/{topic_id}": {
   "get": {
    "summary": "Get Topic",
    "operationId": "get_topic_api_topics__topic_id__get",
    "parameters": [
     {
      "name": "topic_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Topic Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TopicDetailOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/chats/{chat_id}": {
   "get": {
    "summary": "Get Chat",
    "operationId": "get_chat_api_chats__chat_id__get",
    "parameters": [
     {
      "name": "chat_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Chat Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "summary": "Update Chat",
    "description": "Override the classifier: include a chat it skipped (it's analyzed on the next run), or leave one out.",
    "operationId": "update_chat_api_chats__chat_id__patch",
    "parameters": [
     {
      "name": "chat_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Chat Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatSummaryOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/profile/{node_id}": {
   "get": {
    "summary": "Get Profile",
    "operationId": "get_profile_api_profile__node_id__get",
    "parameters": [
     {
      "name": "node_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Node Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProfileOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "summary": "Edit Profile",
    "operationId": "edit_profile_api_profile__node_id__put",
    "parameters": [
     {
      "name": "node_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Node Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileEdit"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProfileEditOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/insights": {
   "get": {
    "summary": "List Insights",
    "operationId": "list_insights_api_insights_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "default": "new",
       "title": "Status"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InsightOut"
         },
         "title": "Response List Insights Api Insights Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/insights/{insight_id}/feedback": {
   "post": {
    "summary": "Post Insight Feedback",
    "operationId": "post_insight_feedback_api_insights__insight_id__feedback_post",
    "parameters": [
     {
      "name": "insight_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Insight Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InsightAction"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/ask": {
   "get": {
    "summary": "Ask History",
    "operationId": "ask_history_api_ask_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AskMessageOut"
         },
         "type": "array",
         "title": "Response Ask History Api Ask Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Ask Question",
    "operationId": "ask_question_api_ask_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AskIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Clear Ask",
    "operationId": "clear_ask_api_ask_delete",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AskIn": {
    "properties": {
     "content": {
      "type": "string",
      "title": "Content"
     }
    },
    "type": "object",
    "required": [
     "content"
    ],
    "title": "AskIn"
   },
...
   "ChatUpdate": {
    "properties": {
     "is_learning": {
      "type": "boolean",
      "title": "Is Learning"
     }
    },
    "type": "object",
    "required": [
     "is_learning"
    ],
    "title": "ChatUpdate"
   },
...
   "Estimate": {
    "properties": {
     "chats": {
      "type": "integer",
      "title": "Chats"
     },
     "turns": {
      "type": "integer",
      "title": "Turns"
     },
     "usd": {
      "type": "number",
      "title": "Usd"
     },
     "breakdown": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Breakdown"
     }
    },
    "type": "object",
    "required": [
     "chats",
     "turns",
     "usd",
     "breakdown"
    ],
    "title": "Estimate"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "NodeOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "is_learning": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Learning"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "parent_id",
     "kind",
     "title",
     "status",
     "is_learning",
     "created_at"
    ],
    "title": "NodeOut"
   },
   "NodeUpdate": {
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "move": {
      "type": "boolean",
      "title": "Move",
      "default": false
     }
    },
    "type": "object",
    "title": "NodeUpdate"
   },
...
   "ProfileEdit": {
    "properties": {
     "markdown": {
      "type": "string",
      "title": "Markdown"
     }
    },
    "type": "object",
    "required": [
     "markdown"
    ],
    "title": "ProfileEdit"
   },
...
   "RunOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "stage": {
      "type": "string",
      "title": "Stage"
     },
     "stage_done": {
      "type": "integer",
      "title": "Stage Done"
     },
     "stage_total": {
      "type": "integer",
      "title": "Stage Total"
     },
     "chats": {
      "type": "integer",
      "title": "Chats"
     },
     "counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Counts"
     },
     "errors": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Errors"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "status",
     "stage",
     "stage_done",
     "stage_total",
     "chats",
     "counts",
     "errors",
     "error",
     "created_at",
     "updated_at"
    ],
    "title": "RunOut"
   },
...
   "StartAnalysis": {
    "properties": {
     "limit": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit"
     }
    },
    "type": "object",
    "title": "StartAnalysis"
   },
...
   "TopicDetailOut": {
    "properties": {
     "topic": {
      "$ref": "#/components/schemas/TopicOut"
     },
     "summary": {
      "type": "string",
      "title": "Summary"
     },
     "breadcrumbs": {
      "items": {
       "$ref": "#/components/schemas/Crumb"
      },
      "type": "array",
      "title": "Breadcrumbs"
     },
     "children": {
      "items": {
       "$ref": "#/components/schemas/TopicOut"
      },
      "type": "array",
      "title": "Children"
     },
     "patterns": {
      "items": {
       "$ref": "#/components/schemas/PatternOut"
      },
      "type": "array",
      "title": "Patterns"
     },
     "stickies": {
      "items": {
       "$ref": "#/components/schemas/StickyOut"
      },
      "type": "array",
      "title": "Stickies"
     },
     "chats": {
      "items": {
       "$ref": "#/components/schemas/ChatSummaryOut"
      },
      "type": "array",
      "title": "Chats"
     }
    },
    "type": "object",
    "required": [
     "topic",
     "summary",
     "breadcrumbs",
     "children",
     "patterns",
     "stickies",
     "chats"
    ],
    "title": "TopicDetailOut"
   },
...
  }
 }
}